requirements.txt
langchain
chromadb
streamlit
google-generativeai
langchain-google-genai
python-dotenv
pypdf